# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GDial Emergency Notification System API",
    "version": "0.1.0"
  },
  "paths": {
    "/settings/system": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get All System Settings",
        "description": "Get all system settings.",
        "operationId": "get_all_system_settings_settings_system_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SystemSettingResponse"
                  },
                  "type": "array",
                  "title": "Response Get All System Settings Settings System Get"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Bulk Update System Settings",
        "description": "Bulk update system settings.",
        "operationId": "bulk_update_system_settings_settings_system_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemSettingBulkUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Bulk Update System Settings Settings System Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Create System Setting",
        "description": "Create a new system setting.",
        "operationId": "create_system_setting_settings_system_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemSettingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemSettingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/settings/system/{key}": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get System Setting",
        "description": "Get a system setting by key.",
        "operationId": "get_system_setting_settings_system__key__get",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemSettingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Update System Setting",
        "description": "Update a system setting.",
        "operationId": "update_system_setting_settings_system__key__put",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemSettingUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemSettingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "settings"
        ],
        "summary": "Delete System Setting",
        "description": "Delete a system setting.",
        "operationId": "delete_system_setting_settings_system__key__delete",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemSettingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/settings/system/groups": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get Settings Groups",
        "description": "Get settings organized by group.",
        "operationId": "get_settings_groups_settings_system_groups_get",
        "parameters": [
          {
            "name": "group",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Group"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SettingsGroup"
                  },
                  "title": "Response Get Settings Groups Settings System Groups Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/settings/dtmf": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get Dtmf Settings Api",
        "description": "Get DTMF settings.",
        "operationId": "get_dtmf_settings_api_settings_dtmf_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DtmfSettingResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Update Dtmf Settings",
        "description": "Update DTMF settings.",
        "operationId": "update_dtmf_settings_settings_dtmf_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DtmfSettingUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DtmfSettingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/settings/sms": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get Sms Settings Api",
        "description": "Get SMS settings.",
        "operationId": "get_sms_settings_api_settings_sms_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SmsSettingsResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Update Sms Settings",
        "description": "Update SMS settings.",
        "operationId": "update_sms_settings_settings_sms_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmsSettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SmsSettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/settings/notifications": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get Notification Settings Api",
        "description": "Get notification settings.",
        "operationId": "get_notification_settings_api_settings_notifications_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationSettingsResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Update Notification Settings",
        "description": "Update notification settings.",
        "operationId": "update_notification_settings_settings_notifications_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationSettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationSettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/settings/security": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get Security Settings Api",
        "description": "Get security settings.",
        "operationId": "get_security_settings_api_settings_security_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecuritySettingsResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Update Security Settings",
        "description": "Update security settings.",
        "operationId": "update_security_settings_settings_security_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SecuritySettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecuritySettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DtmfSettingResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid4",
            "title": "Id"
          },
          "max_attempts": {
            "type": "integer",
            "title": "Max Attempts"
          },
          "input_timeout": {
            "type": "integer",
            "title": "Input Timeout"
          },
          "confirm_response": {
            "type": "boolean",
            "title": "Confirm Response"
          },
          "retry_on_invalid": {
            "type": "boolean",
            "title": "Retry On Invalid"
          },
          "additional_digits": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Additional Digits"
          },
          "universal_gather": {
            "type": "boolean",
            "title": "Universal Gather"
          },
          "repeat_message_digit": {
            "type": "string",
            "title": "Repeat Message Digit"
          },
          "confirm_receipt_digit": {
            "type": "string",
            "title": "Confirm Receipt Digit"
          },
          "request_callback_digit": {
            "type": "string",
            "title": "Request Callback Digit"
          },
          "transfer_to_live_agent_digit": {
            "type": "string",
            "title": "Transfer To Live Agent Digit"
          },
          "dtmf_menu_style": {
            "type": "string",
            "title": "Dtmf Menu Style"
          },
          "inter_digit_timeout": {
            "type": "integer",
            "title": "Inter Digit Timeout"
          },
          "allow_message_skip": {
            "type": "boolean",
            "title": "Allow Message Skip"
          },
          "extra_settings": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extra Settings"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "max_attempts",
          "input_timeout",
          "confirm_response",
          "retry_on_invalid",
          "universal_gather",
          "repeat_message_digit",
          "confirm_receipt_digit",
          "request_callback_digit",
          "transfer_to_live_agent_digit",
          "dtmf_menu_style",
          "inter_digit_timeout",
          "allow_message_skip",
          "created_at",
          "updated_at"
        ],
        "title": "DtmfSettingResponse",
        "description": "Response schema for DTMF settings."
      },
      "DtmfSettingUpdate": {
        "properties": {
          "max_attempts": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Attempts"
          },
          "input_timeout": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Timeout"
          },
          "confirm_response": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirm Response"
          },
          "retry_on_invalid": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry On Invalid"
          },
          "additional_digits": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Additional Digits"
          },
          "universal_gather": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Universal Gather"
          },
          "repeat_message_digit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repeat Message Digit"
          },
          "confirm_receipt_digit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirm Receipt Digit"
          },
          "request_callback_digit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Callback Digit"
          },
          "transfer_to_live_agent_digit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transfer To Live Agent Digit"
          },
          "dtmf_menu_style": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dtmf Menu Style"
          },
          "inter_digit_timeout": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inter Digit Timeout"
          },
          "allow_message_skip": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allow Message Skip"
          },
          "extra_settings": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extra Settings"
          }
        },
        "type": "object",
        "title": "DtmfSettingUpdate",
        "description": "Schema for updating DTMF settings."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SettingsGroup": {
        "properties": {
          "group_name": {
            "type": "string",
            "title": "Group Name"
          },
          "settings": {
            "items": {
              "$ref": "#/components/schemas/SystemSettingResponse"
            },
            "type": "array",
            "title": "Settings"
          }
        },
        "type": "object",
        "required": [
          "group_name",
          "settings"
        ],
        "title": "SettingsGroup",
        "description": "Schema for a group of settings."
      },
...
      "SystemSettingBulkUpdate": {
        "properties": {
          "settings": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Settings"
          }
        },
        "type": "object",
        "required": [
          "settings"
        ],
        "title": "SystemSettingBulkUpdate",
        "description": "Schema for bulk updating system settings."
      },
      "SystemSettingCreate": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "value_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value Type",
            "default": "string"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group",
            "default": "general"
          }
        },
        "type": "object",
        "required": [
          "key",
          "value",
          "description"
        ],
        "title": "SystemSettingCreate",
        "description": "Schema for creating a new system setting."
      },
      "SystemSettingResponse": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "value_type": {
            "type": "string",
            "title": "Value Type"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "group": {
            "type": "string",
            "title": "Group"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "key",
          "value",
          "value_type",
          "description",
          "group",
          "created_at",
          "updated_at"
        ],
        "title": "SystemSettingResponse",
        "description": "Response schema for system settings."
      },
      "SystemSettingUpdate": {
        "properties": {
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "value_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value Type"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group"
          }
        },
        "type": "object",
        "title": "SystemSettingUpdate",
        "description": "Schema for updating a system setting."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}